+++ R/feature_groups-screening.R
@@ -538,7 +538,7 @@ setMethod("filter", "featureGroupsScreening", function(obj, ..., onlyHits = NULL
 #' @aliases screenSuspects
 #' @export
 setMethod("screenSuspects", "featureGroups", function(fGroups, suspects, rtWindow, mzWindow, adduct, skipInvalid,
-                                                      prefCalcChemProps, onlyHits)
+                                                      prefCalcChemProps, neutralChemProps, onlyHits)
 {
     checkmate::assertFlag(skipInvalid) # not in assert collection, should fail before assertSuspectList
 
@@ -548,7 +548,8 @@ setMethod("screenSuspects", "featureGroups", function(fGroups, suspects, rtWindo
     ac <- checkmate::makeAssertCollection()
     assertSuspectList(suspects, needsAdduct = needsAdduct, skipInvalid, add = ac)
     aapply(checkmate::assertNumber, . ~ rtWindow + mzWindow, lower = 0, finite = TRUE, fixed = list(add = ac))
-    aapply(checkmate::assertFlag, . ~ skipInvalid + prefCalcChemProps + onlyHits, fixed = list(add = ac))
+    aapply(checkmate::assertFlag, . ~ skipInvalid + prefCalcChemProps + neutralChemProps + onlyHits,
+           fixed = list(add = ac))
     checkmate::reportAssertions(ac)
 
     if (!is.null(adduct))
@@ -556,9 +557,10 @@ setMethod("screenSuspects", "featureGroups", function(fGroups, suspects, rtWindo
     
     # do this before checking cache to ensure proper errors/warnings are thrown!
     suspects <- prepareSuspectList(suspects, adduct, skipInvalid, checkDesc = TRUE,
-                                   prefCalcChemProps = prefCalcChemProps)
+                                   prefCalcChemProps = prefCalcChemProps, neutralChemProps = neutralChemProps)
     
-    hash <- makeHash(fGroups, suspects, rtWindow, mzWindow, adduct, skipInvalid, prefCalcChemProps, onlyHits)
+    hash <- makeHash(fGroups, suspects, rtWindow, mzWindow, adduct, skipInvalid, prefCalcChemProps, neutralChemProps,
+                     onlyHits)
     cd <- loadCacheData("screenSuspects", hash)
     if (!is.null(cd))
         return(cd)