+++ R/utils-screening.R
@@ -18,9 +18,10 @@ doScreeningShow <- function(obj)
     printf("Suspects annotated: %s\n", if (!is.null(screenInfo(obj)[["estIDLevel"]])) "yes" else "no")
 }
 
-prepareSuspectList <- function(suspects, adduct, skipInvalid, checkDesc, prefCalcChemProps, calcMZs = TRUE)
+prepareSuspectList <- function(suspects, adduct, skipInvalid, checkDesc, prefCalcChemProps, neutralChemProps,
+                               calcMZs = TRUE)
 {
-    hash <- makeHash(suspects, adduct, skipInvalid, calcMZs)
+    hash <- makeHash(suspects, adduct, skipInvalid, checkDesc, prefCalcChemProps, neutralChemProps, calcMZs)
     cd <- loadCacheData("screenSuspectsPrepList", hash)
     if (!is.null(cd))
         suspects <- cd
@@ -65,7 +66,8 @@ prepareSuspectList <- function(suspects, adduct, skipInvalid, checkDesc, prefCal
         }
         
         if (checkDesc)
-            suspects <- prepareChemTable(suspects, prefCalcChemProps = prefCalcChemProps)
+            suspects <- prepareChemTable(suspects, prefCalcChemProps = prefCalcChemProps,
+                                         neutralChemProps = neutralChemProps)
         
         # calculate ionic masses if possible (not possible if no adducts are given and fGroups are annotated)
         if (calcMZs && (is.null(suspects[["mz"]]) || any(is.na(suspects[["mz"]]))) &&