+++ NEWS.md
@@ -51,6 +51,8 @@ Active logins are now necessary to use webservices such as CSI:FingerID, see e.g
 - Fixed: `screenSuspects()` would fail if the adduct column contains partially `NA` data.
 - Fixed: filtered sets data (peak lists/annotations) could sometimes lead to errors
 - `annotateSuspects()`: Multiple conditions for ID level estimation can now be combined with the `and` keyword in the `YAML` configuration file. This is especially useful when combined with the \code{or} keyword.
+- Validation of formula data in e.g. suspect lists is now much faster when `prefCalcChemProps=FALSE`
+- Chemical data from e.g. suspects and TPs can now be 'neutralized' by setting the `neutralChemProps`/`neutralizeTPs` arguments.
 
 
 # patRoon 2.1